refactor: 글쓰기 모달 게시판 목록 조회 위치 수정
