+++ apps/web/src/entities/feed/index.ts
@@ -1,3 +1,4 @@
 export * from './api';
 export * from './config';
 export * from './model';
+export * from './ui';
